update test cases and add response entity

+++ src/main/java/com/ria/springweb/controllers/BirdRestController.java
@@ -2,36 +2,65 @@ package com.ria.springweb.controllers;
 
 import com.ria.springweb.entities.Bird;
 import com.ria.springweb.service.BirdService;
-import org.springframework.stereotype.Controller;
+import jakarta.validation.Valid;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
+import java.text.DateFormat;
+import java.text.SimpleDateFormat;
+import java.util.Calendar;
+import java.util.Date;
 import java.util.List;
+import java.util.stream.Collectors;
 
 @RestController
 public class BirdRestController {
-    BirdService birdService = new BirdService();
+
+    @Autowired
+    BirdService birdService;
     @RequestMapping(value = "/birds", method = RequestMethod.GET)
-    public List<Bird> getBirds(){
-        return birdService.getBirds();
+    public ResponseEntity<List<Bird>> getBirds(){
+        if(birdService.getBirds().isEmpty()) return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        return new ResponseEntity<>(birdService.getBirds().stream().filter(Bird::isVisible).collect(Collectors.toList()),
+                HttpStatus.OK);
     }
 
     @RequestMapping(value = "/birds/{id}", method = RequestMethod.GET)
-    public Bird getBird(@PathVariable("id") int id){
-        return birdService.getBird(id);
+    public ResponseEntity<Bird> getBird(@PathVariable("id") int id){
+        if(birdService.getBird(id)==null) return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        return new ResponseEntity<>(birdService.getBird(id), HttpStatus.OK);
     }
 
     @RequestMapping(value = "/birds", method = RequestMethod.POST)
-    public Bird addBird(@RequestBody Bird bird){ return birdService.addBird(bird);
+    public ResponseEntity<Bird> addBird( @RequestBody @Valid Bird bird){
+        if(bird.getAdded()==null){
+            Date date = Calendar.getInstance().getTime();
+            DateFormat dateFormat = new SimpleDateFormat("yyyy-mm-dd");
+            bird.setAdded(dateFormat.format(date));
+        }
+        return new ResponseEntity<>(birdService.addBird(bird), HttpStatus.CREATED);
     }
-
     @RequestMapping(value = "/birds", method = RequestMethod.PUT)
-    public Bird updateBird(@RequestBody Bird bird){
-        return birdService.updateBird(bird);
+    public ResponseEntity<Bird> updateBird(@RequestBody @Valid Bird bird){
+        if(bird.getAdded()==null){
+            Date date = Calendar.getInstance().getTime();
+            DateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd");
+            bird.setAdded(dateFormat.format(date));
+        }
+
+        return new ResponseEntity<>(birdService.updateBird(bird), HttpStatus.CREATED);
     }
 
     @RequestMapping(value = "/birds/{id}", method = RequestMethod.DELETE)
-    public void deleteBird(@PathVariable("id") int id){
-        birdService.deleteBird(id);
+    public ResponseEntity<Bird> deleteBird(@PathVariable("id") int id){
+        if(birdService.getBird(id)==null) return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        else{
+            birdService.deleteBird(id);
+            return ResponseEntity.status(HttpStatus.OK).build();
+        }
     }
 
 }
